requirements.txt
django
pandas
pdfplumber
langchain
langchain-community
pinecone-client
openai
chardet
tiktoken
python-slugify